refactor(catalog): move Application CreateCategory to vertical slice folder
