+++ src/Ecommerce.Catalog/Ecommerce.Catalog.Application/Categories/CreateCategory/CreateCategoryCommand.cs
@@ -1,5 +1,5 @@
 using Ecommerce.Shared.Application.Commands;
 
-namespace Ecommerce.Catalog.Application.Categories.Commands;
+namespace Ecommerce.Catalog.Application.Categories.CreateCategory;
 
 public sealed record CreateCategoryCommand(string Name, string? Description) : ICommand<int>;
